+++ src/common/py/integration/authorization/strategies/basic/basic_strategy.py
@@ -116,9 +116,9 @@ def create_basic_strategy(
     auth_token: AuthorizationToken | None = None,
     auth_public: AuthorizationSettings | None = None,
     auth_private: AuthorizationSettings | None = None,
-) -> BasicStrategy:
+) -> AuthorizationStrategy:
     """
-    Creates a new OAuth2 strategy instance.
+    Creates a new Basic strategy instance.
 
     Args:
         comp: The main component.